test: add seed structure to project on dev environment
