+++ package.json
@@ -16,7 +16,6 @@
     "services:down": "docker compose -f ./tests/infra/docker-compose.yaml down  ",
     "migrate:up": "prisma migrate dev",
     "migrate:down": "prisma migrate reset",
-    "prisma:studio": "prisma studio",
     "test": "vitest run --config ./tests/setup/vitest.config.mts",
     "test:watch": "vitest --config ./tests/setup/vitest.config.mts",
     "test:cov": "vitest --coverage",
@@ -61,6 +60,7 @@
     "ts-loader": "9.5.2",
     "ts-node": "10.9.2",
     "tsconfig-paths": "4.2.0",
+    "tsx": "^4.23.0",
     "typescript": "5.7.3",
     "typescript-eslint": "8.20.0",
     "unplugin-swc": "1.5.9",
