+++ prisma.config.ts
@@ -7,5 +7,6 @@ export default defineConfig({
   },
   migrations: {
     path: 'prisma/migrations',
+    seed: 'tsx prisma/seed.ts',
   },
 });
